Add files via upload
This is The Complete Papers Explaining the How and the means and the Implications Of this Mathematical Construction. Please Note that this is just the beginning, The ramifications implied are exponential. The uses of this findings will have an Impact on 1. The "Algorithmization" of Algebraic Geometry, 2. A "Grand Unified Theory" of Math and Physics, 3. Revolution in Cryptography and Error Correction, 4. Advanced Computational Complexity (P vs NP), 5. Finality of the "Transcendental Gap"...and beyond.